Added most of the HTML
I added most of the HTML elements using JavaScript, not all are in, such as the buttons for the carousel, however I will be adding those later on top of creating the CSS and the carousel functionality. I also plan on adding a dark mode and a hover effect on all of my buttons.